Add files via upload
Generates TF-IDF word clouds for top 10 words by time period. Colors are defined manually. If a different number of top words/scores (e.g. 20) is imported the script should be easy to modify, but if a different number of categories (e.g. number of genres), the script and color definition must be manually updated.